Fix windows build
Fix the windows build. Problem was the VLA in blkmax_inverted_index.cc
nop compiling under MSVC++.

Also hnsw.saveIndex wasn't checking if the file could be opened - which
was masking a problem in the unittest. Fix the unittest here, and
also branch out to test a couple more methods.