add a ParallelFor function
Add a ParallelFor function with the goal of removing openmp
as a dependency.

This enables multi-threaded indexing for HNSW on OSX - as well
as fixes a problem in the python bindings with knnQueryBatch
where if a thread threw an exception the exception wouldn't
be propagated to the main thread (fix as suggested in issue #178).