Allow Python Handling of log messages
This change lets Python LogHandlers handle the log messages
produced by nmslib.

This lets python control the behaviour
of the log messages at run time, meaning that the loglevel
can be adjusted and that the output of the messages can
be handled any python loghandler.

This basically involves transforming the existing logger class into
an abstract base, and then creating a python version that adapts
the nmslib log system to the python version.