travis CI changes
OSX build are taking too long to finish, mark these builds as optional (allow_failure), and then
set the 'fast_finish' flag so that we report the build as passing as soon as the linux builds
finish.

Also test out python3.6 on linux too, and build extras on linux as well.